MT-01-41

+++ src/components/MyCities/MyCities.jsx
@@ -3,6 +3,7 @@ import { useSelector, useDispatch } from 'react-redux'
 import toDoActions from '../../redux/actions/toDoActions'
 import MyCitiesCard from '../MyCitiesCard.jsx/MyCitiesCard'
 import { useState } from 'react'
+import Swal from 'sweetalert2'
 import './mycities.css'
 
 function MyCities() {
@@ -10,7 +11,7 @@ function MyCities() {
   let { getCitiesUser, getAndDestroy } = toDoActions
   const dispatch = useDispatch()
 
-  const { cities } = useSelector((state)=> state.cities)
+  const { citiesAdmin } = useSelector((state)=> state.cities)
 
   const [userId, setUserId] = useState('')
 
@@ -20,13 +21,33 @@ function MyCities() {
 
   let listenDeleted=(id, e)=>{
 
-    console.log(id)
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Deleted!',
+          'Your city has been deleted.',
+          'success'
+        )
+        console.log(id)
     
-    setId(id)
-    dispatch(getAndDestroy({cityId: id}))
-    
-    alert('ciudad destruidisima') 
-    dispatch(getCitiesUser({userId:userId})) 
+        dispatch(getAndDestroy({cityId: id}))
+
+        dispatch(getCitiesUser({userId:userId}))
+        if (userId.length !== 24){
+          alert('el admind id es invalido')
+          dispatch(getCitiesUser())
+        }
+      }
+    })
+
   }
 
 
@@ -49,7 +70,7 @@ function MyCities() {
     </button>  
     </div>
     <div className='containerCardsHotel'>
-      {cities.map(e=><MyCitiesCard key={e._id} event1={listenDeleted} id={e._id} name={e.title} img={e.image} ></MyCitiesCard>)}
+      {citiesAdmin.map(e=><MyCitiesCard key={e._id} event1={listenDeleted} id={e._id} name={e.title} img={e.image} ></MyCitiesCard>)}
     </div>
     
     </>
